requirements.txt
aniso8601==7.0.0
asgiref==3.3.0
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-filter==2.4.0
django-reversion==3.0.8
djangorestframework==3.12.2
docutils==0.16
drf-writable-nested==0.6.2
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.10
promise==2.3
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2020.4
requests==2.25.0
Rx==1.6.1
sentry-sdk==0.19.4
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Unidecode==1.1.1
urllib3==1.26.2
whitenoise==5.2.0
